Add RenderTarget trait.
Updated RenderWindow and RenderTexture to use it,
and appropriate imports where needed.  Should probably
update the Drawable trait now, as that's where the most
benefit will show up.

This is a minorly breaking change as it needs an extra import
and possibly some changes where views for RenderWindows are needed.