Changed to the new design of c_str
Rust was changed to use a new design on CString. This commit make sure
that rust-sfml use this new design.

This now compiles using the newest nightly, and every example runs. I
have not tested everything.

I am not sure that everything is 100% correct. But at least it compiles
and the examples work.