Remove rust-ci integration
Rust-ci support has been broken for a long time, and no working
solution was found.

Moreover, the rust-ci service has been down for a long time as well,
with no news of being put back online.

This commit removes the broken rust-ci support, with hopes that we can
integrate with a similar service in the future.