Remove non-functional `Rc` versions of resource-using types
These versions haven't been maintained for a long time, possibly because
they are not very amenable to maintenance. They are essentially copy-pasted
versions of the non-`Rc` versions.

Even if we want to support this feature in the future, we should implement
it in a more maintainable way, possibly using generics. If that's not
possible, macros and other forms of code generation are still a more
maintainable solution than the status quo.