Remove "short example" from documentation
Since it is in the root module, but is using the
`window` and `graphics` modules, it makes the doc-tests fail
if those features are not enabled.

It's not a very compelling example anyway.

The user is already directed towards the `examples` subfolder in
the readme.

Also there could be more examples in the different submodules, if desired.