Make RenderStates copyable
It only consists of copyable types and shared references, so it's safe
to copy.

Furthermore, SFML ends up consuming the RenderStates when drawing with
it. In Rust, this results in not being able to draw with the same
RenderStates more than once, due to move semantics. Making it `Copy` fixes
this.