Rework RenderStates to avoid needless copying
RenderStates is now a repr(transparent) of its ffi counterpart,
and we just do a simple pointer cast when sending it to ffi.

This should reduce the number of copying done during drawing.

A breaking change is that Drawable::draw takes RenderStates by immutable
reference. If the RenderStates needs to be changed, first a copy has
to be made that can then be mutated.