Implement vector operations that are in main SFML Vector2.ini
Here is a link to file we need to match:
https://github.com/SFML/SFML/blob/master/include/SFML/System/Vector2.inl

Mul, and Div impl needed to be changed to match Main SFML a bit more.
Added cwise_mul and cwise_div because that was what Div and Mul
originally inteded to do.

Added dot, perpendicular, length_sq, and cross because they were simple
to implement, and I believe they should be in rust-sfml also.