Implement function to fetch raw texture behind an rc_texture
This can be useful for many things like allowing the user the use
Sprites, Shapes, VertexArrays, etc... with the RcTexture. I found this
limitation out recently and realized I kind of needed this function to
continue working. I don't see and negative side effects because it is
using a borrow, so you can't really use that underlying raw texture as
long as the RcTexture is alive.