Implement safer pixel data functions
I believe that it is unnecessary for the user to use unsafe functions
when it can easily be implemented with safe functionality. This is an
example of doing it safely with a small abstraction. Rust is all about
safe code, not unsafe code. Such a simple abstraction should be hidden
from the user